Moves MapSpace and MapSpaceCollection into Rooms namespace, which is where they are actually used.
