+++ ParquetClassLibrary/Rooms/MapSpace.cs
@@ -2,9 +2,10 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using ParquetClassLibrary.Maps;
+using ParquetClassLibrary.Parquets;
 using ParquetClassLibrary.Utilities;
 
-namespace ParquetClassLibrary.Parquets
+namespace ParquetClassLibrary.Rooms
 {
     /// <summary>
     /// A <see cref="ParquetStack"/> together with its coordinates within a given <see cref="Maps.MapRegion"/>.
